chore: Run npm reset-project and remove script from package.json
- Keeping example code as a reference, though it may not be needed.
